wip: REGULAR 고객은 할인이 없다

+++ implementation-solid-by-tdd/src/main/kotlin/com/currenjin/CheckoutService.kt
@@ -11,4 +11,4 @@ class CheckoutService {
 }
 
 data class Order(val items: List<Item>, val customerTier: String)
-data class Item(val name: String, val price: Double, val qty: Int)
+data class Item(val name: String, val price: Long, val qty: Int)
